+++ packages/app/src/components/familyInformation/familyInformation.js
@@ -2,23 +2,6 @@ import React from "react";
 import styles from "./styles.module.css";
 
 function FamilyMemberInfo(props) {
-  const fields = [
-    {
-      label: "Name",
-      value: props.familyRepresentativeName
-    },
-    {
-      label: "Email Address",
-      value: props.familyRepresentativeEmailAddress
-    },
-    {
-      label: "Phone Number",
-      value: props.familyRepresentativePhoneNumber
-    },
-    {
-      label: "Relationship",
-      value: props.representativeRelationship
-    }];
   return (
     <div className={styles.main}>
       <div className={styles.header}>
@@ -26,7 +9,7 @@ function FamilyMemberInfo(props) {
       </div>
       <div className={styles.content}>
         {
-          fields.map((obj) => (<div key={obj.label}>
+          props.familyinfo.map((obj) => (<div key={obj.label}>
             <label>{obj.label}</label>
             <span>{obj.value}</span>
           </div>))
