+++ packages/app/src/components/nomination/nominationShow.js
@@ -1,6 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import NominationBanner from './nominationBanner'
 import nominationsAPI from '../../utils/API/nominationsAPI';
+import HealthProvider from '../healthProvider/healthProvider';
+import FamilyMemberInfo from '../familyInformation/familyInformation';
+
+const dummyStyle = {
+  margin: '0 auto',
+  backgroundColor: 'var(--light-background)',
+  padding: '2em',
+}
 
 const NominationShow = ({ match: { params: { id } } }) => {
   const [NominationData, setNominationData] = useState({})
@@ -32,10 +40,62 @@ const NominationShow = ({ match: { params: { id } } }) => {
       </div>
     );
   }
+  
+  const fields = [
+    {
+      label: "Name",
+      value: NominationData.providerName
+    },
+    {
+      label: "Email Address",
+      value: NominationData.providerEmailAddress
+    },
+    {
+      label: "Phone Number",
+      value: NominationData.providerPhoneNumber
+    },
+    {
+        label: "Title",
+        value: NominationData.providerTitle
+    },
+    {
+      label: "Email Validated",
+      value: NominationData.emailValidated
+    },
+    {
+      label: "Public Email Domain",
+      value: NominationData.publicEmailDomain
+    },
+    {
+      label: "Patient Diagnosis",
+      value: NominationData.patientDiagnosis
+    }];
 
+    const familyinfo = [
+      {
+        label: "Name",
+        value: NominationData.representativeName
+      },
+      {
+        label: "Email Address",
+        value: NominationData.representativeEmailAddress
+      },
+      {
+        label: "Phone Number",
+        value: NominationData.representativePhoneNumber
+      },
+      {
+        label: "Relationship",
+        value: NominationData.representativeRelationship
+      }];
+    
   return (
     <div className="nomination-show-page">
       <NominationBanner nomination={NominationData && NominationData}/>
+      <div style={dummyStyle}>
+        <HealthProvider fields={fields} />
+        <FamilyMemberInfo familyinfo={familyinfo} />
+      </div>
     </div>
   );
 };
